requirements.txt
joblib>=1.3
matplotlib>=3.7
numpy>=1.24
pillow>=10.0
scikit-learn>=1.3
